backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
selenium==4.16.0
beautifulsoup4==4.12.2
pymongo==4.6.1
dnspython==2.4.2
google-generativeai==0.3.2
python-dotenv==1.0.0
